+++ R/auto-clean.R
@@ -9,7 +9,7 @@
 #' For glycomics data, this function calls these functions in sequence:
 #' - [auto_remove()]
 #' - [auto_impute()]
-#' - [auto_aggregate()]
+#' - [aggregate()]
 #' - [auto_normalize()]
 #' - [auto_correct_batch_effect()]
 #'
@@ -17,7 +17,7 @@
 #' - [auto_remove()]
 #' - [auto_normalize()]
 #' - [auto_impute()]
-#' - [auto_aggregate()]
+#' - [aggregate()]
 #' - [auto_normalize()]
 #' - [auto_correct_batch_effect()]
 #'
@@ -50,7 +50,8 @@
 #' exp <- real_experiment
 #' auto_clean(exp)
 #'
-#' @seealso [auto_normalize()], [auto_remove()], [auto_impute()], [auto_aggregate()], [auto_correct_batch_effect()]
+#' @seealso [auto_normalize()], [auto_remove()], [auto_impute()], [aggregate()],
+#'   [auto_correct_batch_effect()]
 #' @export
 auto_clean <- function(
   exp,
@@ -117,7 +118,7 @@ auto_clean <- function(
   cli::cli_alert_success("Imputation completed.")
 
   cli::cli_h2("Aggregating data")
-  exp <- auto_aggregate(exp, standardize_variable = params$standardize_variable)
+  exp <- aggregate(exp, standardize_variable = params$standardize_variable)
   cli::cli_alert_success("Aggregation completed.")
 
   cli::cli_h2("Normalizing data again")
@@ -157,7 +158,7 @@ auto_clean <- function(
   cli::cli_alert_success("Imputation completed.")
 
   cli::cli_h2("Aggregating data")
-  exp <- auto_aggregate(exp, standardize_variable = params$standardize_variable)
+  exp <- aggregate(exp, standardize_variable = params$standardize_variable)
   cli::cli_alert_success("Aggregation completed.")
 
   cli::cli_h2("Normalizing data")